First refactor, component styled directories
