+++ components/app.js
@@ -1,0 +1,1 @@
+import * as TimerComponent from './timer/timer.js'
