+++ components/timer/styles/timer.css
@@ -1,0 +1,1 @@
+.pie{width:300px;height:300px;display:block;position:relative;border-radius:50%;background-color:#1fbb39;border:2px solid #1fbb39;float:left;margin:2em}.pie .block{position:absolute;background:#fff;width:280px;height:280px;display:block;border-radius:50%;top:10px;left:10px}#pieTime{font-size:3em;position:absolute;top:35%;left:43%;color:#999}.degree{background-image:linear-gradient(90deg, transparent 50%, white 50%),linear-gradient(90deg, white 50%, transparent 50%)}
